+++ salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/TagDTO.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public class TagDTO implements java.io.Serializable {
 
-	private int idTag;
+	private Long idTag;
 	private String name;
 	private boolean state;
 	private List<ProductTagDTO> productTags = new ArrayList<ProductTagDTO>(0);
@@ -17,24 +17,24 @@ public class TagDTO implements java.io.Serializable {
 	public TagDTO() {
 	}
 
-	public TagDTO(int idTag, String name, boolean state) {
+	public TagDTO(Long idTag, String name, boolean state) {
 		this.idTag = idTag;
 		this.name = name;
 		this.state = state;
 	}
 
-	public TagDTO(int idTag, String name, boolean state, List<ProductTagDTO> productTags) {
+	public TagDTO(Long idTag, String name, boolean state, List<ProductTagDTO> productTags) {
 		this.idTag = idTag;
 		this.name = name;
 		this.state = state;
 		this.productTags = productTags;
 	}
 
-	public int getIdTag() {
+	public Long getIdTag() {
 		return this.idTag;
 	}
 
-	public void setIdTag(int idTag) {
+	public void setIdTag(Long idTag) {
 		this.idTag = idTag;
 	}
